Bug#17378106 - VALGRIND FAILURE IN TWO RPL TESTS / WEEKLY RUN
In the implementation of row_scan_index_for_mysql() which is called from
handler::records() and handler::check()a buffer is allocated with size
UNIV_PAGE_SIZE for row_search_for_mysql(). Then a mysql-formatted buffer
is put into it.  If the TABLE_SHARE::record length is > UNIV PAGE SIZE,
this valgrind assert could occur in 5.7 due to the call of
handler::records() for any count(*) because ::records() presets
prebuilt->mysql_row_len to the actual TABLE_SHARE record length.

This is only an issue with a valgrind assert.  There is never an actual
buffer overrun because row_scan_index_for_mysql() never reads of-page
columns, so the record read is never actually longer than UNIV_PAGE_SIZE.

The fix is simply to allocate the buffer to receive the record as the
maximum of UNIV_PAGE_SIZE and prebuilt->mysql_row_len.

Approved by Marko in rb3286