Bug #17441867 [ERROR] INNODB: FAILED TO SET O_DIRECT ON FILE
PROBLEM
-------

Setting the O_DIRECT flag on file on tmpfs fails 
and error messages are printed in error logs.
The customer was creating lot of temporary tables 
in tmpfs and because of which error log was 
flooded with these error messages. We however 
ignore these errors and continue to create the
tables.These error messages were causing problem 
with the their monitoring system.

FIX
---

The fix is to downgrade the error with a warning
and printing this warning only once in the error
log.