BUG#16985214 - VALGRIND ERRORS, INCONSISTENT RESULT IN STORED PROCEDURE
BACKGROUND:
When mysqld is started with valgrind and we run the minimal 
testcase provided in the bug report (In a Stored Procedure,
we declare a variable of type geometry, copy the value 
contained in the table to this variable and then just display 
this variable with SELECT statement on the console) 
valgrind complains of reading freed memory.

ANALYSIS:
Looking at the valgrind stacktrace, while sending the data
back to the client (in SELECT e statement) valgrind complains
of reading freed memory.
After some more analysis, it is found that actual problem is 
happening in SELECT z into e from t1; statement
After analyzing the call flow the problem happens while
copying with memmove() where it copies from->ptr to to->ptr

How blobs are represented in Memory?
For field class we have ptr variable which stores the pointer
to the actual data. However for blobs, it stores the
packlength in first 4 bytes and in rest 8 bytes it stores
the starting address of the memory where blob is actually
stored in memory.
NOTE: Packlength is the number of bytes used to represnt blobs
in memory.

from->ptr
----------------------------	      ------------------ 	
|packlength| blob pointer   |-------> |Actual blob data|
----------------------------          ------------------

With memmove we are just copying the from->ptr to to->ptr.
Basically we are doing a "shallow copy". to->ptr is also 
pointing to the same temporary memory location where from->ptr
is pointing.
How copying with memmove looks in memory

from->ptr
----------------------------	      ------------------ 	
|packlength| blob pointer---|-------> |Actual blob data|
----------------------------          -------^----------
				             |
to->ptr					     |
----------------------------	      	     |
|packlength| blob pointer---|---------------->
----------------------------          

Later on after select z into e from t1 statement is finished 
executing it calls ha_innobase::reset() where it frees
the temporary heap memory allocated by calling mem_area_free().
After that in SP when we do 3rd statement select e statement
we get valgrind error, Invalid read of freed memory.
This is the root cause of this bug.

How we prevent this valgrind error? 

We prevent this valgrind error by doing a "deep copy" of blobs.
Blob class has a member called String value. We allocate 
a new block of memory to hold the data pointed to by 
blob pointer and copy the starting address to 
value->Ptr.

from->ptr
--------------------------	    ------------------ 	
|packlength| blob pointer |-------> |Actual blob data|
--------------------------          ------------------

We first take out blob pointer from from->ptr by calling
from->val_str() function. 
Then in Field_blob::store it calls Field_geom::store_internal
which calls value.copy().
This function allocates a new block of memory and copies the data
pointed to by blob pointer and returns the starting address of
this memory location and puts it in value->Ptr.

String value        
------------------           -----------------------
| blob pointer1---|-------> | Copy Actual blob data |
------------------	    | from blob pointer     |
value->Ptr		     -----------------------


Till now this is how it looks like in memory.

Then we copy this blob pointer 1 to to->ptr.
This is how it looks like in memory. 

to->ptr
----------------------------	      	
|packlength| blob pointer1--|----------> 
-----------------^----------           |
                 |                     |
	         |                     |
String value     |                     V
----------------------          -----------------------
| blob pointer 1------|-------> |New Actual blob data |
----------------------          -----------------------
value->Ptr

Now both value->Ptr and to->ptr containg the starting address 
of newly allocated blob data
Basically we do a "deep copy" of blobs.
Now we have two copies of blob data 
1. blob pointer which is stored in from->ptr and is pointing
to actual blob data.
2. As a result of deep copy, blob pointer1 which is stored in
to->Ptr and is pointing to the same actual blob data in memory.

WHY VALGRIND WARNING DOES NOT COME OUTSIDE OF SP FOR THESE TWO CASES
THOUGH THEY ALSO INVOKE field_conv() function?

Case 1: CREATE TABLE t2 AS SELECT z FROM t1;

Case 2: CREATE TABLE t3(zz POINT);
        INSERT INTO t3(zz) SELECT z FROM t1; 

To understand why valgrind warning does not come for these two
cases, we have to first understand how 

SELECT z from t1 works.(t1 refers to the table created in
                        minimal testcase in bug report)

When we issue the above query, storage engine innodb has a persistent
pointer named data which is pointing to actual data of blob in memory.
In row_sel_store_mysql_field_func() function in innodb,
1. It allocates a temporary heap for blob
2. It copies the data from persistent memory location pointed to 
   by data to this heap.
3. It displays the data to the user by calling 
   Protocol::send_result_set_row ()which calls 
   Protocol_text::store() and so on.
4. When this is finished executing, it calls close_thread_table() 
   which in turn calls ha_innobase::reset() which calls
   if (prebuilt->blob_heap) {
                row_mysql_prebuilt_free_blob_heap(prebuilt);
        }

Then it frees the temporary heap created for blobs by calling 
mem_heap_free() function.

For Case 1: 
Although this statement also invokes field_conv() function
and copies the data using memmove() (basically shallow copy), 
this data is first made persistent on disk by calling 
ha_innobase::write_row() and it is inserted into the table.

Afterwards the temporary heap allocated for blobs in 
SELECT z from t1 is freed by calling mem_area_free().
Therefore afterwards, when we do SELECT z from t2, since the data 
is already made persistent by storage engine by the above statement
we do not get valgrind warning of invalid read of freed memory.
NOTE: SELECT z from t2 also follows the same working steps described
for SELECT z from t1.

Same reasoning applies to Case 2 INSERT INTO t3(zz) SELECT z FROM t1;
statement. First data is made persistent by Innodb with write_record()
which calls ha_innobase::write_row() etc,
then it frees the temporary heap created for blobs with
in SELECT z FROM t1 statement.

FIX:
As a fix for this Valgrind bug, a small check has been added
in field_conv() function to prevent the copy using memmove().
The callflow is changed to prevent the shallow copy using 
memmove () and do a deep copy of Geometry datatype as it is 
already done for blobs. This is because Since Geometry is also
represented in memory as blob type since it is a subclass of 
Field_blob.
It is also verified that the user-reported problem of inconsistent
results in the initial part of bug report was triggered by the 
incorrect copy method. With the fix, it is working correctly.

NOTE:

1. The patch for this bug also fixes Bug#17339396.
The root cause of the problem mentioned in that bug is the
same as root cause and analysis done in this bug.
Additional test coverage is added for Bug#17339396.
It is verified that patch for Bug#16985214 fixes the problem
mentioned in Bug#17339396. Bug#17339396 can be marked 
as duplicate of Bug#16985214 and can be closed.

2. This valgrind bug happens only with Geometry datatype
in Stored Procedure. Without SP, if we do like this
SELECT z into @e from t1;
SELECT @e;
It does not happen, since it follows the correct call
sequence and calls the right method to copy the values.
Basically does a deep copy. This is also verified.

3. Also it is verified that it does not happen with any other 
datatype inside or outside of SP.