Bug#17503130 TABLE PERFORMANCE_SCHEMA.TABLE_HANDLES BECOMES EMPTY AFTER A BIG WORKLOAD
Before this fix, the implementation of table performance_schema.table_handle
would appear empty after the server executed some code,
causing spurious failures in test cases like mdl_func.

The failure was spurious, because the same test did work after a server
restart.

The root cause is that the logic to check for locks in
table_table_handles::make_row() is broken, mismatched locks are used.

It works only by accident after a server restart,
because the PFS_table::m_lock and the PFS_table_share::m_lock happen to be
initialized to the same initial value.

The fix is to use the proper lock for end_optimistic_lock().