This is a fix for Bug #17435744 'JAM EVENTS ARE SOMETIMES WRITTEN TO THE TRACE FILE OF THE WRONG THREAD'.
In the methods that were affected by this problem, 
jam() calls have been replaced by either jamNoBlock(),
or by thrjam(<jam buffer>), where <jam buffer> is passed
as an argument to the method.
 
This ensures that the jam event is written into the jam buffer of the
calling thread, rather than the thread that the block instance is assigned 
to.  
 
jamNoBlock() reads thread local data (via pthread_getspecific()). To 
avoid the cost of that, a pointer to the jam buffer is passed as a 
function argument in most cases.
 
The patch also adds an assert to thrjamLine(). This assert checks that the
jam buffer belongs to the calling thread. This should prevent similar errors
from being introduced in the future.
 
As an added benefit, this patch will also serve to document which
block member functions that are called by "foreign" threads.

Unrelated cleanup: Unused field Dbtup_client::m_block is removed.