BUG#17458273: KEY LENGTHS LEAD TO INCOMPLETE RESULT SETS;               UNHINTED SCANS PREFER SUCH INDEXES
Consider a table with an index over a prefix of column 'x' and
a query with a range condition over 'x':

CREATE TABLE (... INDEX (x(4)))
SELECT ... WHERE 'foobar2' < x and x < 'foobar5';

The index will not be able to distinguish between rows with 
x-values like "foobar", "foobaz" or "foobar3" since only the 
first four characters ("foob") will be stored in the index.

Since all rows with an x-value prefix "foob" look equal, the
range optimizer has to translate '<' and '>' to '<=' and '>=',
respectively, for such indexes.

In this bug, only "'foobar2' < x" is correctly translated. 
Later, when key_and() is executed on the the two ranges
"'foob' <= x" and "NULL < x < 'foob'", key_and() correctly
finds that these ranges together form an impossible condition.
The fix is to also translate '<' to '<=' for prefix indexes
so that key_and() can ultimately create the range 
"'foob' <= x <= 'foob'".