Fix for Bug#15907122
INCORRECT HANDLING OF JOB-BUFFERS ALMOST FULL - 'SLEEPLOOP 10'

The internal mt-job scheduler (mt.cpp) has the function 
update_sched_config() which is supposed to calculate how
many signals we are allowed to execute in the next
scheduling round. Allowed number of signals are restricted by
available job-buffers (aka 'signal buffers') for the 
block-threads we may communicate with.

If the available job-buffers for *any* such thread are below 
a critical threshold, update_sched_config() will wait for more
job buffers to become available. The wait is implemented with the
yield() function which takes a 'thr_wait*' argument (Think of this 
as a condition). A 'wakeup(thr_wait*)' (from another thread) 
will later restart the suspended thread.

However, the 'thr_wait*' argument used in these yield() / wakeup()
was the one used for signal that *incomming* signals has become
available. This is the opposite of what we want to wait for. The
effect of this was that these yields waited the maximum timeout (1ms)
before execution was resumed.

This fix introduce an additional 'thr_wait*' which is signaled when
job buffers are released and thus becomes available.

This fix is a collection of the following sub patches:

 
 - Add dump of job buffer utilization before we
   crash due to 'job buffer full'

 - Change the signature of the yield() function such that it can
   take a more general '*check_callback()' function as arguments.

   Change the existing functions currently used as 'callbacks'
   to the new (relaxed) signature.

-  Refactor: Splitt out compute_free_buffers_in_queue()
   from compute_max_signals_to_execute().
   New function contains common code intended for reuse in later
   patches.

- (Fix) update_sched_config() will 'yield' the CPU and wait for
   more job buffers to become available when it is about
   to run out of job buffers.

   The yield() call will wait on a 'thr_wait' object, which
   may be sent a 'wakeup()' when the waiting condition has
   been resolved (by another thread)

   However, update_sched_config() yield'ed on the incorrect
   'thr_wait' object intended to be used to wait for more
   *incomming signal* - What we actually have to wait for
   are signals to be *consumed* by the destination thread.
   Luckily there is also defined a 'max wait' of 1ms which
   currently will wakeup the thread. Thus, we never waited
   forever, but added latency could be expected.

   This patch introduce a new 'thr_wait' object which is
   signaled by the consumer, and update_sched_config()
   will now wait on this object.

   Furthermore the patch also avoid a situation where a
   thread could end up waiting for itself.