Bug#17539520 PERFORMANCE_SCHEMA THREAD CREATION PERFORMS TOO MUCH INITIALIZATION
This fix is a performance improvement.

A lot of code present in create_thread() initialized the per thread internal
buffers that support:
- table events_waits_current
- table events_stages_current
- table events_statements_current

Executing this code to fully initialize every member for every row is in
fact un necessary, and the code has been removed.

Only the initialization of the default NESTING_EVENT_ID and
NESTING_EVENT_TYPE for the thread waits and stages is required,
and the overall initialization of these columns has been cleaned up.

As a related fix,
PFS_events_waits::m_thread was redundant with PFS_events::m_thread_internal_id,
and has been removed.

Logic in table_events_waits_common::make_row() that was in fact specific to
only two sub classes has been moved to
- table_events_waits_current::make_row()
- table_events_waits_history::make_row()