WL#6606: Offload THD initialization and network initialization          to worker thread
Follow-up patch.

This patch removes the "end_thread_signal" signal handler.
It was currently unused. It was designed to allow sigint to be
sent to individual threads in debug mode. The problem with
the signal handler was that it used unsafe functions.

The patch also includes several code simplifications made possible
by the removal of "end_thread_signal".