Bug#17310878 PFS OVERHEAD ON FREQUENT CONNECT/DISCONNECT
This fix is a back port from 5.7 to 5.6

This fix is a performance improvement.

The issue was that the call to PSI_THREAD_CALL(delete_current_thread)(),
which is executed each time a thread disconnects,
was placed inside a critical section involving LOCK_thread_count.

The fix is to perform the same call sooner,
before entering the LOCK_thread_count critical section.

Serializing all calls to pfs_delete_current_thread_v1() was the major
cause of performance overhead, for frequent connect/disconnect.