BUG#11748566 - MYSQL SHOWS WRONG DEFAULT VALUE FOR PK IF                NOT NULL NOT SPECIFIED
DESCRIPTION:

Two different, but equivalent ways of declaring a column
as primary key does not end-up with the same result. When the
column is declared primary key inline as:
    CREATE TABLE t (a INT PRIMARY KEY)
it has no default value. In this case the flag
NO_DEFAULT_VALUE_FLAG for the column is set while parsing.
On the other hand, in the second case when column is declared
primary key as index:
    CREATE TABLE t (a INT, PRIMARY KEY(a))
the flag is not set. This results in a default value for
the primary key column.

FIX:

When a column is declared as primary key as index, the flag
NO_DEFAULT_VALUE_FLAG is set for that column if user has not
provided an explicit default value for the column. For the
TIMESTAMP/DATETIME fields, if explicit_defaults_for_timestamp
variable is not set, default value assigned due to first
timestamp column promotion is retained. Default constant value
assigned due to implicit promotion of second timestamp column
is removed.