Bug#17434385: SHOW PROCEDURE CODE ASSERTION FAILED: AT SQL_LEX.CC:2831
After WL#6369 all explainable statements are supposed to be not cleaned till
the end of the statement in order to provide plan for EXPLAIN FOR CONNECTION.
To ensure that an assert at sql_lex.cc:2831 was added. SHOW PROCEDURE CODE
can't be explained, so it's ok to skip this assert.
Also, "<already_cleaned_up>" isn't printed on a clena unit since it's essentially
useless: if unit is clean all select list items are reverted to initial state and
don't point to a tmp table anymore, so it's safe to print them.