Bug #16630410 NDB:LCP FRAGMENT SCAN WATCHDOG SHOULD BE CONFIGURABLE
The LCP fragment scan watchdog periodically checks for lack of
progress in a fragment scan performed as part of a local checkpoint.
The watchdog was hard-coded to shut down the node if there was no
progress for 60 seconds.

Added a configuration parameter LcpScanProgressTimeout to set the
maximum time for which the local checkpoint can be stalled before
the LCP fragment scan watchdog shuts down the node. A parameter
value of 0 disables the LCP fragment scan watchdog.

Added a unit test to set the parameter value to 0 and verify that
the LCP fragment scan watchdog is disabled.