Bug#16960133 MYSQLBINLOG VERBOSE OUTPUT, INVALID MEMORY ACCESS IN MY_B_WRITE_QUOTED
Analysis: IN RBR mode and binlog_row_image is set to
MINIMAL, only the changed columns are logged in binlog.
And the meta data (eg: null bytes) will also be written
only for these effected columns. `mysqlbinlog` tool in
verbose mode (-v) skips these null bytes before reading
the actual row. This tool assumes that for each column
in the table there is one null bit which is not always
true(eg: binlog_row_image=minimal) and calculates the
number of total null bytes written in the binlog wrongly
(ie., skips wrong number of bytes) and finally ends up
in interpreting the row in wrong way. This problem could
lead to invalid memory reads. The formula used to calculate
the metadata bytes was ((number of columns in the table + 7) /8).

For example, if we have 9 columns in a table and if there
is a two columns insert/update event, master would have
written two bits (i.e., one byte) in row metadata. If
we calculate based on number of columns in table, it 
results 2 bytes instead of 1 byte.

Fix:
The formula should be based on the number of columns
written in the event instead of the total number of columns
in the table. Changed new formula is
((number of columns written in the event) + 7)/8)