wl#7279 : Add engine clause to main suite subtest which requires myisam
Few tests are failing (when run with innodb as default engine) because one/more
subtests from ".test" file are myisam specific. We can't skip run of entire test with
innodb as only small part of test is myisam.
We added engine=myisam clause to create table statement where subtest is myisam specific.
Hence it will make subtest run with myisam and rest of test with default engine.
Result file modified with related to this change and for additional expected difference.