Bug#16894092: PERFORMANCE REGRESSION IN 5.6.6+ FOR INSERT               INTO .... SELECT ... FROM
Problem:
The performance regression is related to writing the temporary table
into MyISAM. If we configure tmp_table_size and max_heap_table_size
sufficiently high (in this case setting them to 96M) in order to ensure that
the temporary table can be stored in the Memory engine, the performance
regression is no longer present.
Regression source was found to be WL#6071.

Analysis:
While creating temporary table, a random scan is initialized.
This initialization internally frees buffers in case of myisam.
As a result writes/reads happen to disk rather than
cache. And hence the performance hit when we try to create
tmp table in myisam.

Solution:
Initialize the scan without freeing buffers which can be done
by calling ha_rnd_init(1).