Bug #17581965: MYSQL-TRUNK FAILS TO COMPILE ON VS2013RC
The fix for bug #17332056 was incomplete as it didn't bump up the minimum windows version required 
through _WIN32_WINNT.

Fixed by updating the CMAKE windows script to require the actual minimal version.
Also added a check for VS2013RC implementing isnan in math.h before emabrking on 
the workaround to define is as a synonim for _isnan().

No test case needed