Bug#11752837 NDB_MGM -E " START BACKUP WAIT STARTED" NOT WORKS AS DOCUMENTED
When a 'start backup wait started' command was executed on the 
command-line using the '-e' option of ndb_mgm, the client did not 
exit until the backup was completed.

Modified ndb_mgm for a 'wait started' backup so that it exits 
immediately after printing 'Backup started' instead of waiting for 
backup completion.